feat: refactor avatar service to use Avatar class and update documentation
