+++ Classes/Image/Avatar.php
@@ -2,19 +2,21 @@
 
 declare(strict_types=1);
 
-namespace KonradMichalik\Typo3LetterAvatar\Utility;
+namespace KonradMichalik\Typo3LetterAvatar\Image;
 
 use KonradMichalik\Typo3LetterAvatar\Enum\ImageDriver;
 use KonradMichalik\Typo3LetterAvatar\Image\Driver\GdAvatar;
 use KonradMichalik\Typo3LetterAvatar\Image\Driver\ImagickAvatar;
-use KonradMichalik\Typo3LetterAvatar\Image\LetterAvatarInterface;
+use KonradMichalik\Typo3LetterAvatar\Utility\ConfigurationUtility;
 use TYPO3\CMS\Core\Utility\Exception\NotImplementedMethodException;
 
-class ImageDriverUtility
+class Avatar
 {
-    public static function resolveAvatarService(...$args): LetterAvatarInterface
+    public static function create(...$args): LetterAvatarInterface
     {
-        switch (ConfigurationUtility::get('imageDriver', ImageDriver::class)) {
+        $imageDriver = $args['imageDriver'] ?? ConfigurationUtility::get('imageDriver', ImageDriver::class);
+
+        switch ($imageDriver) {
             case ImageDriver::IMAGICK:
             default:
                 return new ImagickAvatar(...$args);
